Change zone configuration from/to catalog from/to any other type:
When changing to "Producer" replication type, RFC-recommended `invalid.`
namerserver and `soa` from the `catalog`  config section are applied.

When changing from "Producer" to "Master" or "Native" replication type, the
default nameservers template and the the specified SOA are applied.

Zone content (all records except SOA and apex NS RRSet) is kept as this
does not cause issue to PowerDNS: when a zone is set to "Producer", all
non-catalog records are not served.